significant rewrite to separate out the UI components from the dtcyber engine proper.  right now this is unix only -- windows parts not done yet, probably broken for the moment
git-svn-id: svn://akdesign.dyndns.org/dtcyber/trunk@295 d8f3d347-7192-4e35-a67b-6d701703d080

+++ doelz.c
@@ -185,8 +185,8 @@ static void *doelzThread(void *param);
 **  Public Variables
 **  ----------------
 */
-u16 doelzPort;
-u16 doelzConns;
+long doelzPort;
+long doelzConns;
 int doelzNodeCnt;
 DoelzNode *doelzNodes;
